chore: file.c copy method move for linux in the separate file
