+++ core/src/main/c/share/files.c
@@ -35,7 +35,6 @@
 #else
 
 #include <utime.h>
-#include <sys/sendfile.h>
 
 #endif
 
@@ -110,35 +109,6 @@ JNIEXPORT jint JNICALL Java_io_questdb_std_Files_close0
         (JNIEnv *e, jclass cl, jlong fd) {
     return close((int) fd);
 }
-
-#ifndef __APPLE__
-JNIEXPORT jint JNICALL Java_io_questdb_std_Files_copy
-        (JNIEnv *e, jclass cls, jlong lpszFrom, jlong lpszTo) {
-    const char* from = (const char *) lpszFrom;
-    const char* to = (const char *) lpszTo;
-    const int input = open(from, O_RDONLY);
-    if (-1 ==  (input)) {
-        return -1;
-    }
-
-    const int output = creat(to, 0644);
-    if (-1 == (output)) {
-        close(input);
-        return -1;
-    }
-
-    // On linux sendfile can accept file as well as sockets
-    off_t offset = 0;
-    struct stat fileStat = {0};
-    fstat(input, &fileStat);
-    int result = sendfile(output, input, &offset, fileStat.st_size);
-
-    close(input);
-    close(output);
-
-    return result;
-}
-#endif
 
 JNIEXPORT jlong JNICALL Java_io_questdb_std_Files_openRW
         (JNIEnv *e, jclass cl, jlong lpszName) {
